+++ tests/test-0745av1_single_frame.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_745av1_single_frame
+describe "mkvmerge / AV1 OBUs with only a single frame"
+test_merge "data/av1/single_frame.obu"